refact: rearrange constants
