+++ software/mobile/screens/trips/trip-query/view-constants.tsx
@@ -19,12 +19,12 @@ export enum ItemTypeKey {
  */
 export interface DropdownItem {
   label: string;
-  value: metricConstants.ServerMetricKey | string;
+  value: ServerMetricKey | string;
   parent: string | undefined;
 }
 
 /**
- * Types
+ * Types and Enums
  */
 export const allowedStatistics = [
   "avg_vibration",
@@ -38,6 +38,20 @@ export type StatisticsMeasurementPacketKey =
   | "avg_noise"
   | "avg_temperature"
   | "avg_velocity";
+
+/**
+ * The server metric key
+ *
+ * The server uses snake casing for the database.
+ */
+export enum ServerMetricKey {
+  VIBRATION_METRIC_KEY = "avg_vibration",
+  NOISE_METRIC_KEY = "avg_noise",
+  TEMPERATURE_METRIC_KEY = "avg_temperature",
+  VELOCITY_METRIC_KEY = "avg_velocity",
+  END_DATE_KEY = "end_time_s",
+  START_DATE_KEY = "start_time_s",
+}
 
 /**
  * Text
@@ -76,13 +90,13 @@ export const TRIP_PROPERTY_ITEMS: DropdownItem[] = [
 
   {
     label: metricConstants.START_DATE_LABEL,
-    value: metricConstants.ServerMetricKey.START_DATE_KEY,
+    value: ServerMetricKey.START_DATE_KEY,
     parent: ItemTypeKey.Date,
   },
 
   {
     label: metricConstants.END_DATE_LABEL,
-    value: metricConstants.ServerMetricKey.END_DATE_KEY,
+    value: ServerMetricKey.END_DATE_KEY,
     parent: ItemTypeKey.Date,
   },
 
@@ -96,9 +110,7 @@ export const TRIP_PROPERTY_ITEMS: DropdownItem[] = [
     label:
       metricConstants.AVERAGE_LABEL_PREFIX +
       metricConstants.VIBRATION_METRIC_TITLE,
-    value:
-      metricConstants.AVERAGE_KEY_PREFIX +
-      metricConstants.ServerMetricKey.VIBRATION_METRIC_KEY,
+    value: ServerMetricKey.VIBRATION_METRIC_KEY,
     parent: ItemTypeKey.Number,
   },
 
@@ -106,18 +118,14 @@ export const TRIP_PROPERTY_ITEMS: DropdownItem[] = [
     label:
       metricConstants.AVERAGE_LABEL_PREFIX +
       metricConstants.TEMPERATURE_METRIC_TITLE,
-    value:
-      metricConstants.AVERAGE_KEY_PREFIX +
-      metricConstants.ServerMetricKey.TEMPERATURE_METRIC_KEY,
+    value: ServerMetricKey.TEMPERATURE_METRIC_KEY,
     parent: ItemTypeKey.Number,
   },
 
   {
     label:
       metricConstants.AVERAGE_LABEL_PREFIX + metricConstants.NOISE_METRIC_TITLE,
-    value:
-      metricConstants.AVERAGE_KEY_PREFIX +
-      metricConstants.ServerMetricKey.NOISE_METRIC_KEY,
+    value: ServerMetricKey.NOISE_METRIC_KEY,
     parent: ItemTypeKey.Number,
   },
 
@@ -125,9 +133,7 @@ export const TRIP_PROPERTY_ITEMS: DropdownItem[] = [
     label:
       metricConstants.AVERAGE_LABEL_PREFIX +
       metricConstants.VELOCITY_METRIC_TITLE,
-    value:
-      metricConstants.AVERAGE_KEY_PREFIX +
-      metricConstants.ServerMetricKey.VELOCITY_METRIC_KEY,
+    value: ServerMetricKey.VELOCITY_METRIC_KEY,
     parent: ItemTypeKey.Number,
   },
 ];
